+++ pkg/artifactory/resource/replication/resource_artifactory_pull_replication_test.go
@@ -2,12 +2,13 @@ package replication_test
 
 import (
 	"fmt"
-	"github.com/jfrog/terraform-provider-shared/util"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"regexp"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/acctest"
+	"github.com/jfrog/terraform-provider-shared/test"
+	"github.com/jfrog/terraform-provider-shared/util"
 )
 
 func mkTclForPullRepConfg(name, cron, url string) string {
@@ -39,7 +40,7 @@ func mkTclForPullRepConfg(name, cron, url string) string {
 
 func TestAccPullReplicationInvalidCron(t *testing.T) {
 
-	_, fqrn, name := acctest.MkNames("lib-local", "artifactory_pull_replication")
+	_, fqrn, name := test.MkNames("lib-local", "artifactory_pull_replication")
 	var failCron = mkTclForPullRepConfg(name, "0 0 * * * !!", acctest.GetArtifactoryUrl(t))
 
 	resource.Test(t, resource.TestCase{
@@ -56,7 +57,7 @@ func TestAccPullReplicationInvalidCron(t *testing.T) {
 }
 
 func TestAccPullReplicationLocalRepo(t *testing.T) {
-	_, fqrn, name := acctest.MkNames("lib-local", "artifactory_pull_replication")
+	_, fqrn, name := test.MkNames("lib-local", "artifactory_pull_replication")
 	config := mkTclForPullRepConfg(name, "0 0 * * * ?", acctest.GetArtifactoryUrl(t))
 	updatedConfig := mkTclForPullRepConfg(name, "1 0 * * * ?", acctest.GetArtifactoryUrl(t))
 	resource.Test(t, resource.TestCase{
@@ -92,8 +93,8 @@ func TestAccPullReplicationLocalRepo(t *testing.T) {
 }
 
 func TestAccPullReplicationRemoteRepo(t *testing.T) {
-	_, fqrn, name := acctest.MkNames("lib-remote", "artifactory_pull_replication")
-	_, fqrepoName, repoName := acctest.MkNames("lib-remote", "artifactory_remote_maven_repository")
+	_, fqrn, name := test.MkNames("lib-remote", "artifactory_pull_replication")
+	_, fqrepoName, repoName := test.MkNames("lib-remote", "artifactory_remote_maven_repository")
 	var tcl = `
 		resource "artifactory_remote_maven_repository" "{{ .remote_name }}" {
 			key 				  = "{{ .remote_name }}"