+++ pkg/artifactory/resource/repository/local/resource_artifactory_local_cargo_repository.go
@@ -7,7 +7,7 @@ import (
 	"github.com/jfrog/terraform-provider-shared/util"
 )
 
-var cargoLocalSchema = util.MergeSchema(
+var cargoLocalSchema = util.MergeMaps(
 	BaseLocalRepoSchema,
 	map[string]*schema.Schema{
 		"anonymous_access": {