+++ pkg/artifactory/resource/repository/local/resource_artifactory_local_rpm_repository.go
@@ -9,7 +9,7 @@ import (
 	"github.com/jfrog/terraform-provider-shared/validator"
 )
 
-var rpmLocalSchema = util.MergeSchema(
+var rpmLocalSchema = util.MergeMaps(
 	BaseLocalRepoSchema,
 	map[string]*schema.Schema{
 		"yum_root_depth": {