+++ pkg/artifactory/resource/repository/remote/resource_artifactory_remote_go_repository.go
@@ -16,7 +16,7 @@ type GoRemoteRepo struct {
 func ResourceArtifactoryRemoteGoRepository() *schema.Resource {
 	const packageType = "go"
 
-	var goRemoteSchema = util.MergeSchema(BaseRemoteRepoSchema, map[string]*schema.Schema{
+	var goRemoteSchema = util.MergeMaps(BaseRemoteRepoSchema, map[string]*schema.Schema{
 		"vcs_git_provider": {
 			Type:             schema.TypeString,
 			Optional:         true,