+++ pkg/artifactory/resource/security/resource_artifactory_permission_target_test.go
@@ -2,13 +2,14 @@ package security_test
 
 import (
 	"fmt"
-	"github.com/jfrog/terraform-provider-shared/util"
-	"testing"
-
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
+	"testing"
+
 	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/acctest"
 	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/artifactory/resource/security"
+	"github.com/jfrog/terraform-provider-shared/test"
+	"github.com/jfrog/terraform-provider-shared/util"
 )
 
 const permissionNoIncludes = `
@@ -141,9 +142,9 @@ const permissionFull = `
 `
 
 func TestAccPermissionTarget_GitHubIssue126(t *testing.T) {
-	_, permFqrn, permName := acctest.MkNames("test-perm", "artifactory_permission_target")
-	_, _, repoName := acctest.MkNames("test-perm-repo", "artifactory_local_generic_repository")
-	_, _, username := acctest.MkNames("artifactory_user", "artifactory_user")
+	_, permFqrn, permName := test.MkNames("test-perm", "artifactory_permission_target")
+	_, _, repoName := test.MkNames("test-perm-repo", "artifactory_local_generic_repository")
+	_, _, username := test.MkNames("artifactory_user", "artifactory_user")
 	testConfig := `
 		resource "artifactory_local_generic_repository" "{{ .repo_name }}" {
 		  key             = "{{ .repo_name }}"
@@ -200,7 +201,7 @@ func TestAccPermissionTarget_GitHubIssue126(t *testing.T) {
 }
 
 func TestAccPermissionTarget_full(t *testing.T) {
-	_, permFqrn, permName := acctest.MkNames("test-perm", "artifactory_permission_target")
+	_, permFqrn, permName := test.MkNames("test-perm", "artifactory_permission_target")
 
 	tempStruct := map[string]string{
 		"repo_name":       "example-repo-local",
@@ -283,7 +284,7 @@ func TestAccPermissionTarget_user_permissions(t *testing.T) {
 }
 
 func TestAccPermissionTarget_addBuild(t *testing.T) {
-	_, permFqrn, permName := acctest.MkNames("test-perm", "artifactory_permission_target")
+	_, permFqrn, permName := test.MkNames("test-perm", "artifactory_permission_target")
 
 	tempStruct := map[string]string{
 		"repo_name":       "example-repo-local", // because of race conditions in artifactory, this repo must first exist